[WFCORE-834] Split CapabilityContext impls out of ModelControllerImpl
